demo2-geo-platform/postgis-ingest/requirements.txt
paho-mqtt==1.6.1
psycopg[binary]==3.1.18
python-json-logger==2.0.7
